Add azure desktop dashboard script and tidy up scripts in root

+++ src/scripts/pscore/tear-down-local.ps1
@@ -1,1 +1,9 @@
-helm delete bc -n buyingcatalogue
+$context=kubectl config current-context
+
+if ($context -ne "docker-desktop") {
+    Write-Host "$context is not a local context!"
+    exit 1
+}
+else {
+    helm delete bc -n buyingcatalogue
+}
